aws: Config use custom types instead of primitives [Breaking Change]
Switches all Config primitive values to custom types to ensure configurations
can be merged correctly between service instantiation without oddities such as
not being able to set a value to its Zero because the Zero value is used to
determine if the incoming configuration parameter is set.

The custom types contain the state of if the value was set vs unset. This allows
the Config to be merged correctly in cases such as DiableSSL which previous
could never be disabled via a Merge().

To facilitate this change and to clarify what aws.String function does it was
renamed to aws.StringPtr. This applied to the other pointer helper functions
as well. An IntPtr was also added for *int.

    String  -> StringPtr
    Boolean -> BoolPtr
            -> IntPtr     // New
    Long    -> Int64Ptr
    Double  -> Float64Ptr
    Time    -> TimePtr

In addition helper functions were added for each type to return the value that
a pointer points to or the type's zero value if the pointer is nil. This saves
some of the hassle of checking if a field is nil before dereferencing it in
circumstances where nil should be treated as zero.

The custom types defined in aws/types.go provide wrappers for string, bool, int,
int64, and float64. Each wrapper provides a function to Set() or Get() the
wrapped value. The IsSet() method can be used to determine if the value was
set at all. Get() will return the type's zero value if the wrapper was not
previously set. This does mean also that any place the Config value was read
a .Get() needs to be appended to each value's usage.

Example:
    // Code Prior to this update
    svc := s3.New(&aws.Config{
        Region:           "us-west-2",
        MaxRetries:       2,
        S3ForcePathStyle: true,
    })
    // Prior code read Config.Region value
    if svc.Config.Region == "us-west-2" {...}

    // After Updating to this change:
    svc := s3.New(&aws.Config{
        Region:           aws.String("us-west-2"),
        MaxRetries:       aws.Int(2),
        S3ForcePathStyle: aws.Bool(true),
    })
    // After this change reading Config.Region value
    if svc.Config.Region.Get() == "us-west-2" {...}

Fixes #157